Loggin / meio Footer

+++ application/controllers/Jornalista.php
@@ -5,16 +5,50 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 class Jornalista extends CI_Controller {
 
     public function __construct() {
-        //chama o construtos da classe pai (CI_Controller)
         parent::__construct();
-        //chama o método que faz a validação de login de usuario
         $this->load->model('Jornalista_model');
-        $this->load->view('Header/Header');
-        
     }
 
     public function index() {
-        $this->lista();
+        $this->login();
+    }
+
+    public function login() {
+        $this->form_validation->set_rules('email', 'email', 'required');
+        $this->form_validation->set_rules('senha', 'senha', 'required');
+
+        if ($this->form_validation->run() == false) {
+            $this->load->view('Jornalista/Login');
+        } else {
+
+            $usuario = $this->Jornalista_model->getUsuario(
+                    $this->input->post('email'), $this->input->post('senha')
+            );
+            //valida se retornoun algum registro, quer dizer que o usuário é existente
+            if ($usuario) {
+                $data = array(
+                    'idUsuario' => $usuario->id,
+                    'email' => $usuario->email,
+                    'logado' => TRUE,
+                    'admin' => $usuario->admin
+                );
+                //mandamos o CodeIgniter salvar na sessão os dados do usuario percebva quie o metodo set_userdata é diferente do método set_flashdata,
+                //pois ele slva dados permanentes enquanto durar a sessão
+                $this->session->set_userdata($data);
+                //abre a pagina principal(padrão) do sistema
+                redirect($this->config->base_url());
+            } else {
+                $this->session->set_flashdata('mensagem', 'Usuário e Senha incorretos.');
+                //redireciona para a pagina de login OBRIGANDO fazer o login.
+                redirect($this->config->base_url() . 'Jornalista/login');
+            }
+        }
+    }
+
+    public function sair() {
+        //apaga todo o conteudo da sessão do usuario
+        $this->session->sess_destroy();
+        redirect($this->config->base_url() . 'Jornalista/Login');
     }
 
     public function lista() {
@@ -28,6 +62,7 @@ class Jornalista extends CI_Controller {
         $data['jornalistas'] = $this->Jornalista_model->getAll();
 
         //chama a view passando o conteudo listado (getAll=buscar todos) da variavel $data (variavel que se refere ao banco de dados)
+        $this->load->view('Header/Header');
         $this->load->view('Jornalista/Lista', $data);
         $this->load->view('Footer/Footer');
     }
@@ -37,6 +72,7 @@ class Jornalista extends CI_Controller {
     }
 
     public function cadastrar() {
+
         //fazendo a validação
         $this->form_validation->set_rules('nome', 'nome', 'required'); //nome do campo, id do campo, se é requirido ou não
         $this->form_validation->set_rules('email', 'email', 'required');
@@ -46,6 +82,7 @@ class Jornalista extends CI_Controller {
         if ($this->form_validation->run() == false) {
 
             $data['jornalistas'] = $this->Jornalista_model->getAll();
+            $this->load->view('Header/Header');
             $this->load->view('Jornalista/Cadastro', $data);
             $this->load->view('Footer/Footer');
         } else {
@@ -73,6 +110,10 @@ class Jornalista extends CI_Controller {
     }
 
     public function alterar($id) {
+
+
+        $this->Jornalista_model->verificaLogin();
+
         if ($id > 0) {
 
             //regras de validação
@@ -85,6 +126,7 @@ class Jornalista extends CI_Controller {
                 //monta uma variavel ($data) para mandar dados para a view e chama o metodo getOnde(pegar 1) do Prova_model
                 //para resgatar os dados da prova a ser alterada
                 $data['jornalista'] = $this->Jornalista_model->getOne($id);
+                $this->load->view('Header/Header');
                 $this->load->view('Jornalista/Altera', $data); //carrega a view do formulario
                 $this->load->view('Footer/Footer');
             } else {
@@ -109,15 +151,21 @@ class Jornalista extends CI_Controller {
     }
 
     public function deletar($id) {
-        if ($id > 0) {
+        if ($this->session->userdata('admin') == '1') {
+            $this->Jornalista_model->verificaLogin();
 
-            //manda para o model deletar e ja valida o retorno para saber se funcionou
-            if ($this->Jornalista_model->delete($id)) {
-                $this->session->set_flashdata('mensagem', '<div class="alert alert-success">Sucesso ao deletar.</div>');
+            if ($id > 0) {
+
+                //manda para o model deletar e ja valida o retorno para saber se funcionou
+                if ($this->Jornalista_model->delete($id)) {
+                    $this->session->set_flashdata('mensagem', '<div class="alert alert-success">Sucesso ao deletar.</div>');
+                } else {
+                    $this->session->set_flashdata('mensagem', '<div class="alert alert-danger>Falha ao deletar.</div>');
+                }
             } else {
                 $this->session->set_flashdata('mensagem', '<div class="alert alert-danger>Falha ao deletar.</div>');
+                redirect('Jornalista/lista');
             }
-            redirect('Jornalista/lista');
         }
     }
